Prepare TagContext
Based on investigation into the existing project,
a CodexCollection can store its own AllTags and RootTags.
In practice, all view models inherit ViewModelBase which
exposes one public static data store that contains the active
CodexCollection, and thus the tag collection.

AllTags and RootTags are not overwritten outside of the CodexCollection
itself, and there only during the rehydration of data.
However these collections are modified to contains new tags as
necessary, though not overwritten.

- Creates TagContext which just reads the current tags collections from
  ViewModelBase.CurrentCollection

Once de/serialization is re-implemented, it should be possible to swap
Tag for Tagv2.